Renamed passkey to publicPrivateKey as it was confusing Closes #298
