+++ src/background/service/flowns.ts
@@ -1,7 +1,7 @@
 import * as secp from '@noble/secp256k1';
 import * as fcl from '@onflow/fcl';
 
-import { signMessageHash } from '@/background/utils/modules/passkey';
+import { signMessageHash } from '@/background/utils/modules/publicPrivateKey';
 import wallet from 'background/controller/wallet';
 import { keyringService, openapiService } from 'background/service';
 import { createPersistStore } from 'background/utils';
